requirements.txt
Django>=4.1
djangorestframework>=3.14.0
drf-yasg>=1.21.7
psycopg2-binary>=2.9.3
pytest-django>=4.5.2
pytest>=7.0.1
django-environ
gunicorn
